+++ beamterm-core/src/gl/ubo.rs
@@ -46,7 +46,7 @@ impl UniformBufferObject {
         Ok(())
     }
 
-    pub fn upload_data<T>(&self, gl: &glow::Context, data: &T) {
+    pub fn upload_data<T: Copy>(&self, gl: &glow::Context, data: &T) {
         self.bind(gl);
         unsafe { buffer_upload_struct(gl, glow::UNIFORM_BUFFER, data, glow::STATIC_DRAW) };
         self.unbind(gl);